fix(haetae): make range-proven MSVC conversions explicit
Resolve the HAETAE C4244/C4245/C4267 sites whose bounds or uint32_t bit patterns are already proven. Preserve the rANS reciprocal bit patterns, validate bounded length/nonce narrowing, and align helper/count types with their actual ranges.

Deliberately leave three areas for deeper review rather than hiding them with casts:
- src/DigitalSignature/HAETAE/haetae_fft.c::_mulrnd16: prove fixed-point FFT magnitude bounds across initialization, complex multiplication, and all butterfly stages before making the int64_t-to-int32_t narrowing explicit.
- src/DigitalSignature/HAETAE/haetae_reduce.c::crypto_haetae_freeze and nearby reduction helpers: audit arithmetic right shifts of potentially negative signed values and replace implementation-defined shift assumptions with portable floor/sign operations before treating the final narrowing as resolved.
- src/DigitalSignature/HAETAE/haetae_poly.c::crypto_haetae_poly_decomposed_pack/unpack: verify the decompose_z1 low-bit range and define the signed-byte wire decode portably as a pair before making the int-to-uint8_t conversion explicit.

Linux HAETAE build/tests passed on the patch branch. Windows MSVC /W4, Ubuntu, and macOS compilation succeeded in PR CI before merge; the full slow cross-platform KAT run was still in progress at merge time.

+++ src/DigitalSignature/HAETAE/haetae_rans_byte.h
@@ -240,7 +240,7 @@ static inline void RansEncSymbolInit(RansEncSymbol* s, uint32_t start, uint32_t 
 
         s->rcp_freq = (uint32_t)(
             ((UINT64_C(1) << (shift + 31)) + freq - 1) / freq);
-        s->rcp_shift = shift - 1;
+        s->rcp_shift = (uint16_t)(shift - 1u);
 
         // With these values, 'q' is the correct quotient, so we
         // have bias=start.
